perf(webapp): stackedBarChart componetization
