+++ webapp/src/components/StackedBarChartReport/index.js
@@ -1,8 +1,11 @@
 import React, { memo, useCallback, useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import FileSaver from 'file-saver'
-import { Box } from '@mui/system'
 import { makeStyles } from '@mui/styles'
+import { useTranslation } from 'react-i18next'
+import { useCurrentPng } from 'recharts-to-png'
+import TooltipDownload from '@mui/material/Tooltip'
+import DownloadOutlined from '@mui/icons-material/DownloadOutlined'
 import {
   IconButton,
   Typography,
@@ -10,106 +13,11 @@ import {
   Switch,
   FormGroup
 } from '@mui/material'
-import TooltipDownload from '@mui/material/Tooltip'
-import DownloadOutlined from '@mui/icons-material/DownloadOutlined'
-import { useTranslation } from 'react-i18next'
-import { useCurrentPng } from 'recharts-to-png'
-import {
-  ComposedChart,
-  Bar,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-  Legend,
-  ResponsiveContainer,
-  Cell
-} from 'recharts'
-
-import { formatWithThousandSeparator } from '../../utils/format-with-thousand-separator'
 
 import styles from './styles'
+import CustomStackedBarChart from './custom-stackedBarChart'
 
 const useStyles = makeStyles(styles)
-
-const RenderChartLegend = () => {
-  const classes = useStyles()
-  const { t } = useTranslation()
-
-  return (
-    <div className={classes.chartLegent}>
-      <a key={`key-claimed-link-chart`}>
-        <Box className={classes.legentCircle} bgcolor="#19647e" />
-        {t('claimed', { ns: 'generalForm' })}
-      </a>
-      <a key={`key-categorized-link-chart`}>
-        <Box className={classes.legentCircle} bgcolor="#ee964b" />
-        {t('categorized', { ns: 'generalForm' })}
-      </a>
-      <a key={`key-unclaimed-link-chart`}>
-        <Box className={classes.legentCircle} bgcolor="#f4d35e" />
-        {t('unclaimed', { ns: 'generalForm' })}
-      </a>
-      <a key={`key-uncategorized-link-chart`}>
-        <Box className={classes.legentCircle} bgcolor="#28afb0" />
-        {t('uncategorized', { ns: 'generalForm' })}
-      </a>
-    </div>
-  )
-}
-
-const CustomTooltip = ({ payload = [], label = '', coinType = '' }) => {
-  const { t } = useTranslation()
-  label = label + ''
-  const arrayLabel = label.split(' ')
-
-  return (
-    <div>
-      <strong>
-        {`${t(arrayLabel[0].toLocaleLowerCase(), { ns: 'generalForm' })}`}
-      </strong>
-      {payload &&
-        payload.map((data, i) => (
-          <div key={`${i}-tooltip`}>
-            <div>
-              {`${
-                data.payload.category
-                  ? t(
-                      (
-                        data.dataKey.split('_')[1] + data.payload.category
-                      ).toLocaleLowerCase(),
-                      {
-                        ns: 'generalForm'
-                      }
-                    )
-                  : t(data.dataKey.split('_')[1].toLocaleLowerCase(), {
-                      ns: 'generalForm'
-                    })
-              }: ${formatWithThousandSeparator(
-                data.payload[data.dataKey],
-                4
-              )} ${coinType}`}
-            </div>
-            <div>
-              {i === 0 &&
-                data.payload?.date &&
-                `${
-                  data.payload.date ? t('date', { ns: 'generalForm' }) : ''
-                }: ${
-                  data.payload.date ? data.payload.date.split('T')[0] : ''
-                } `}
-            </div>
-          </div>
-        ))}
-    </div>
-  )
-}
-
-CustomTooltip.propTypes = {
-  payload: PropTypes.array,
-  label: PropTypes.any,
-  coinType: PropTypes.string
-}
 
 const StackedBarChartReport = ({
   data,
@@ -162,52 +70,12 @@ const StackedBarChartReport = ({
           </FormGroup>
         </div>
       </div>
-
-      <ResponsiveContainer width="100%" height={400}>
-        <ComposedChart
-          height={400}
-          data={data}
-          margin={{
-            top: 24
-          }}
-          ref={barRef}
-        >
-          <CartesianGrid stroke="#f5f5f5" />
-          <XAxis
-            tick={{ fontSize: 10, stroke: '#000', strokeWidth: 0.5 }}
-            dataKey="type"
-            scale="auto"
-          />
-          <YAxis tick={{ fontSize: 10, stroke: '#000', strokeWidth: 0.1 }} />
-          <Tooltip
-            wrapperStyle={{
-              outline: 'none',
-              borderRadius: '4px',
-              backgroundColor: '#F9F9F9',
-              fontSize: '14px',
-              padding: '8px'
-            }}
-            content={<CustomTooltip coinType={coinType} />}
-          />
-          {showLegend && <Legend content={<RenderChartLegend />} />}
-          <Bar dataKey={`${coinType}_`} barSize={88} stackId="a">
-            {data.map(({ type }) => (
-              <Cell
-                key={`cell-${type}`}
-                fill={type === 'income' ? '#19647e' : '#ee964b'}
-              />
-            ))}
-          </Bar>
-          <Bar dataKey={`${coinType}_UN`} barSize={88} stackId="a">
-            {data.map(({ type }) => (
-              <Cell
-                key={`cell-${type}`}
-                fill={type === 'income' ? '#f4d35e' : '#28afb0'}
-              />
-            ))}
-          </Bar>
-        </ComposedChart>
-      </ResponsiveContainer>
+      <CustomStackedBarChart
+        data={data}
+        barRef={barRef}
+        coinType={coinType}
+        showLegend={showLegend}
+      />
     </div>
   )
 }
